Remove select current word function
Rationale: Expand Region package already accomplishes this task by
default (if point is in a word, it selects it first). This also has
the advantage of being able to successively mark an ever larger or
smaller region with single successive key presses. Not as nice as
Vim's `viw`, but we'll take what we can get.